Refactor commands into sub package
commands/base.py
commands/compute_data_disk.py
commands/compute_endpoint.py
commands/compute_image.py
commands/compute_request.py
commands/compute_reserved_ip.py
commands/compute_shell.py
commands/compute_storage.py
commands/compute_vm.py
commands/setup_account.py
commands/storage_account.py
